pyproject.toml
[tool.poetry]
name = "conservision-practice"
version = "0.1.0"
description = ""
authors = ["Luc Bams <lbm@mediaan.com>"]
readme = "README.md"
packages = [{include = "conservision_practice", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
torch = {version = "^2.4.0+cu118", source = "pytorch-gpu"}
torchvision = {version = "^0.19.0+cu118", source = "pytorch-gpu"}
torchaudio = {version = "^2.4.0+cu118", source = "pytorch-gpu"}
lightning = "^2.4.0"
matplotlib = "^3.9.2"
numpy = "1.26.4"
scikit-learn = "^1.5.1"
python-dotenv = "^1.0.1"
albumentations = "^1.4.14"


[[tool.poetry.source]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.2"
tqdm = "^4.66.5"
ipykernel = "^6.29.5"
tensorboard = "^2.17.1"
ipywidgets = "^8.1.5"

[tool.ruff]
extend-include = ["*.ipynb"]

[tool.ruff.format]
quote-style = "single"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
